// src/app/api/shared/enums/pharmacy-benefit-plan-type.enum.ts
export enum PharmacyBenefitPlanType {
    Unknown = 0,
    AidsDrugAssistanceProgram = 1,
    Commercial = 2,
    DurableMedicalEquipmentPlan = 3,
    Government = 4,
    ManagedCareMedicaid = 5,
    MedicareAB = 6,
    MedicareB = 7,
    MedicareD = 8,
    Medicaid = 9,
    StatePharmacyAssistanceProgram = 10,
    StatePlan = 11,
    WorkersComp = 12,
    NoPharmacyBenefits = 13
}